chore: extract bar chart events to a separate file
