+++ src/vis/bar/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './BarSortHook';
 export * from './BarVisStateHook';
+export * from './BarChartMouseEvents';
